fix: Long

+++ backend/challenge/src/main/java/com/hakaton/challenge/api/OrderController.java
@@ -25,7 +25,7 @@ public class OrderController {
     }
 
     @GetMapping(value = "/{id}",produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Order> getOrderById(@PathVariable("id") Integer id) {
+    public ResponseEntity<Order> getOrderById(@PathVariable("id") Long id) {
         Order order = orderService.FetchOrderByIdWithTrades(id);
         return new ResponseEntity<>(order, HttpStatus.OK);
     }
