+++ backend/challenge/src/main/java/com/hakaton/challenge/domain/TradeEntity.java
@@ -20,13 +20,13 @@ public class TradeEntity {
     @SequenceGenerator(name = "tradeSeqGen", sequenceName = "tradeSeqGen", initialValue = 1, allocationSize = 1)
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "tradeSeqGen")
     @Column(name="trade_id", unique=true, nullable=false)
-    private int id;
+    private Long id;
 
     @Column(nullable = false)
-    private int buyOrderId;
+    private Long buyOrderId;
 
     @Column(nullable = false)
-    private int sellOrderId;
+    private Long sellOrderId;
 
     @Column(nullable = false)
     private Date createdDateTime;
